pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "jamuna-bankline-prediction"
version = "0.1.0"
description = "Temporal Fusion Transformer for Jamuna River bankline shift prediction"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.3.0",
    "pytorch-lightning>=2.3.0",
    "pytorch-forecasting>=1.4.0",
    "darts>=0.30.0",
    "optuna>=3.6.0",
    "optuna-integration[pytorch-lightning]>=3.6.0",
    "mlflow>=2.14.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "pandas>=2.2.0",
    "numpy>=2.0.0",
    "scikit-learn>=1.5.0",
    "statsmodels>=0.14.0",
    "ruptures>=1.1.9",
    "openpyxl>=3.1.0",
    "joblib>=1.4.0",
    "loguru>=0.7.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.5.0",
    "mypy>=1.10.0",
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",
]

[project.scripts]
train = "src.training.train:main"
tune = "src.training.tune:main"
serve = "src.serving.api:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["B008"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
